SINGA-155 Remove zookeeper for single-process training
add single-process cluster-runtime implementation.
rearrange functions in cluster_rt.h into:
  * cluster_rt.h - cluster-runtime interface and SPClusterRT impl
  * zk_service.h - zookeeper wrapper and ZKClusterRT impl
  * job_manager.h - job manager
